+++ library/SchemaData/ExternalContent/JsonToSchemaObjects/JsonToSchemaObjectsInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Municipio\SchemaData\ExternalContent\JsonToSchemaObjects;
+
+interface JsonToSchemaObjectsInterface {
+
+    /**
+     * @param string $json
+     * 
+     * @return Municipio\Schema\BaseType[]
+     */
+    public function transform(string $json): array;
+}
