+++ library/SchemaData/ExternalContent/SourceReaders/JsonFileSourceReader.php
@@ -3,7 +3,7 @@
 namespace Municipio\SchemaData\ExternalContent\SourceReaders;
 
 use Municipio\SchemaData\ExternalContent\Filter\SchemaObjectsFilter\SchemaObjectsFilterInterface;
-use Municipio\SchemaData\ExternalContent\JsonToSchemaObjects\JsonToSchemaObjects;
+use Municipio\SchemaData\ExternalContent\JsonToSchemaObjects\JsonToSchemaObjectsInterface;
 use Municipio\SchemaData\ExternalContent\SourceReaders\FileSystem\Contracts\FileExists;
 use Municipio\SchemaData\ExternalContent\SourceReaders\FileSystem\Contracts\FileGetContents;
 
@@ -13,7 +13,7 @@ class JsonFileSourceReader implements SourceReaderInterface
         private string $filePath,
         private SchemaObjectsFilterInterface $schemaObjectsFilter,
         private FileGetContents&FileExists $fileSystem,
-        private JsonToSchemaObjects $jsonToSchemaObjects,
+        private JsonToSchemaObjectsInterface $jsonToSchemaObjects,
     )
     {
     }
